% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartModelMatrix.R
\name{bartModelMatrix}
\alias{bartModelMatrix}
\title{Create a matrix out of a vector or data.frame}
\usage{
bartModelMatrix(
  X,
  numcut = 0L,
  usequants = FALSE,
  type = 7,
  rm.const = FALSE,
  cont = FALSE,
  xinfo = NULL
)
}
\arguments{
\item{X}{A vector or data.frame to create the matrix from.}

\item{numcut}{Integer. The maximum number of cutpoints to consider. If numcut=0, then just 
return a matrix; otherwise, return a list containing a matrix X, a vector numcut and a list xinfo.}

\item{usequants}{Logical. If usequants is FALSE, then the cutpoints in xinfo are generated 
uniformly; otherwise, if TRUE, then quantiles are used for the cutpoints.}

\item{type}{Determines which quantile algorithm is employed.}

\item{rm.const}{Logical. Whether or not to remove constant variables.}

\item{cont}{Logical. Whether or not to assume all variables are continuous.}

\item{xinfo}{You can provide the cutpoints to BART or let BART choose them for you. To provide 
them, use the xinfo argument to specify a list (matrix) where the items (rows) are the 
covariates and the contents of the items (columns) are the cutpoints.}
}
\value{
A list containing the cleaned input data, number of cut points, removed variables, and cut point information.
}
\description{
The external BART functions operate on matrices in memory. Therefore, if the user 
submits a vector or data.frame, then this function converts it to a matrix. Also, it determines 
the number of cutpoints necessary and their values for each column when asked to do so.
}
\keyword{internal}
