% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.ind.R
\name{class.ind}
\alias{class.ind}
\title{Class Indicator Matrix}
\usage{
class.ind(cl)
}
\arguments{
\item{cl}{A factor or vector of class labels.}
}
\value{
A matrix with rows corresponding to observations and columns to classes.
}
\description{
This function creates a class indicator matrix from a factor.
}
