% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.cores.openmp.R
\name{mc.cores.openmp}
\alias{mc.cores.openmp}
\title{OpenMP Cores Count}
\usage{
mc.cores.openmp()
}
\value{
0 if OpenMP is not available, otherwise the number of OpenMP cores.
}
\description{
This function calls the C function to get the number of OpenMP cores available.
}
\keyword{internal}
