% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mebart_cont.R
\name{mebart_cont}
\alias{mebart_cont}
\title{Fit a meBART Model with \emph{Continuous} Output Data}
\usage{
mebart_cont(
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  sparse = FALSE,
  theta = 0,
  omega = 1,
  a = 0.5,
  b = 1,
  augment = FALSE,
  rho = NULL,
  xinfo = matrix(0, 0, 0),
  usequants = FALSE,
  cont = FALSE,
  rm.const = TRUE,
  sigest = NA,
  sigdf = 3,
  sigquant = 0.9,
  k = 2,
  power = 2,
  base = 0.95,
  sigmaf = NA,
  lambda = NA,
  fmean = mean(y.train),
  w = rep(1, length(y.train)),
  ntree = 200L,
  numcut = 100L,
  ndpost = 1000L,
  nskip = 100L,
  keepevery = 1L,
  nkeeptrain = ndpost,
  nkeeptest = ndpost,
  nkeeptestmean = ndpost,
  nkeeptreedraws = ndpost,
  printevery = 100L,
  transposed = FALSE,
  proposal_sigma = meas_error_sigma,
  meas_error_sigma,
  x_mu,
  x_sigma
)
}
\arguments{
\item{x.train}{A matrix or data frame containing the explanatory variables for the training data.}

\item{y.train}{An integer vector containing the binary outcome variable for the training data.}

\item{x.test}{A matrix or data frame of test data (optional, default is an empty matrix). The structure should match `x.train`.}

\item{sparse}{Whether to use Dirichlet additive regression trees (DART).}

\item{theta}{alpha parameter in Linero 2018 equation (2), governs DART prior (DART ONLY).}

\item{omega}{omega parameter for DART.}

\item{a}{Beta prior parameters, used in DART only.}

\item{b}{Beta prior parameters, used in DART only.}

\item{augment}{FALSE uses Assumption 2.2 in Linero 2018, TRUE uses assumption 2.1 (DART ONLY).}

\item{rho}{rho sparsity parameter, used in DART only.}

\item{xinfo}{A matrix of cutpoints (default is an empty matrix). Specifies cutpoints for the model if desired.}

\item{usequants}{If TRUE, uniform quantiles are used for cutpoints; otherwise, uniform cutpoints are used.}

\item{cont}{Specifies if all variables are continuous.}

\item{rm.const}{If TRUE, removes constant columns from `x.train`.}

\item{sigest}{Prior estimate for sigma; if NA, it is estimated from the data.}

\item{sigdf}{Degrees of freedom for the sigma prior.}

\item{sigquant}{Quantile for the sigma prior.}

\item{k}{Prior parameter for the mean of the output data.}

\item{power}{Power parameter for the BART tree-depth prior (beta in original BART paper).}

\item{base}{Base parameter for the BART tree-depth prior (alpha in original BART paper).}

\item{sigmaf}{Parameter for the sigma prior, not scaled by the number of trees.}

\item{lambda}{Scale parameter used in the sigma prior.}

\item{fmean}{A numeric value (default is the mean of `y.train`). The mean of the output variable `y.train`.}

\item{w}{A numeric vector (default is a vector of ones). Weights to multiply the standard deviation. Must be of length n.}

\item{ntree}{The number of trees to use in the model.}

\item{numcut}{The number of cutpoints to consider for each variable.}

\item{ndpost}{The number of posterior draws to return.}

\item{nskip}{The number of observations for burn-in (MCMC).}

\item{keepevery}{The thinning parameter for MCMC.}

\item{nkeeptrain}{Number of training predictions to keep.}

\item{nkeeptest}{Number of test predictions to keep.}

\item{nkeeptestmean}{The number of MCMC iterations to return for the test mean.}

\item{nkeeptreedraws}{Number of tree draws to keep.}

\item{printevery}{How often to print progress during MCMC.}

\item{transposed}{Is the input data transposed? Used if called by `mc.wbart`.}

\item{proposal_sigma}{A numeric value (default is `meas_error_sigma`). Covariance matrix of the proposal distribution.}

\item{meas_error_sigma}{A numeric matrix. Covariance matrix of the measurement error, must be square and match the number of columns in `x.train`.}

\item{x_mu}{A numeric column vector. Mean of the explanatory variables.}

\item{x_sigma}{A numeric matrix. Covariance matrix of the explanatory variables.}
}
\value{
A list with the following elements:
\item{mu}{The mean of the outcome variable `y.train`.}
\item{yhat.train.mean}{The mean of the posterior predictions for the training data.}
\item{yhat.train}{The posterior predictions for the training data.}
\item{yhat.test.mean}{The mean of the posterior predictions for the test data.}
\item{yhat.test}{The posterior predictions for the test data.}
\item{varcount}{An (ndpost * p) matrix of variable counts in the model.}
\item{varprob}{An (ndpost * p) matrix of variable probabilities in the model.}
\item{treedraws}{A list of posterior tree draws.}
\item{x_draws}{Tensor with posterior draws for the training data.}
\item{acceptances}{Whether or not the proposal was accepted.}
\item{proc.time}{The time taken for the C++ code to run.}
\item{varcount.mean}{The mean of the variable counts.}
\item{varprob.mean}{The mean of the variable probabilities.}
\item{rm.const}{A vector specifying which columns were removed from `x.train`.}
}
\description{
This function fits a Bayesian Additive Regression Tree (BART) model which also also models measurement error in the explanatory variables. 
This function specifically handles \emph{continuous} output data.
}
\examples{
# Example usage:
set.seed(0)
x.train <- matrix(rnorm(1000), ncol = 10)
y.train <- rnorm(100)
mdl <- mebart_cont(x.train, y.train,
              ndpost = 500,
              ntree = 50,
              meas_error_sigma = diag(10),
              x_mu = matrix(0, nrow=10, ncol=1),
              x_sigma = diag(10))
preds <- mdl$yhat.train.mean
print(paste0("Training MSE: ", round(mean((y.train - preds)^2), 3)))
}
