% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmebart.R
\name{pmebart}
\alias{pmebart}
\title{Single core BART Predictions}
\usage{
pmebart(
  x.test,
  treedraws,
  mu = 0,
  mc.cores = 1L,
  transposed = FALSE,
  dodraws = TRUE,
  nice = 19L
)
}
\arguments{
\item{x.test}{A matrix of test data.}

\item{treedraws}{A list containing the tree draws.}

\item{mu}{A numeric value to add to the predictions.}

\item{mc.cores}{Number of cores to use for parallel processing.}

\item{transposed}{}

\item{dodraws}{Logical. If TRUE, draws are returned.}

\item{nice}{Nice level for the process.}
}
\value{
A matrix of predictions.
}
\description{
This function performs multicore predictions using BART.
}
