% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mebart.R
\name{predict.mebart}
\alias{predict.mebart}
\title{Predictions for meBART}
\usage{
\method{predict}{mebart}(object, newdata, mc.cores = 1, openmp = (mc.cores.openmp() > 0), ...)
}
\arguments{
\item{object}{A meBART object.}

\item{newdata}{A matrix of new data for prediction.}

\item{mc.cores}{Number of cores to use for parallel processing.}

\item{openmp}{Logical. If TRUE, OpenMP is used for parallel processing.}

\item{...}{Additional arguments.}
}
\value{
A matrix of predictions.
}
\description{
This function makes predictions using a meBART model.
}
\examples{
# Example usage
# Load the meBART package
# library(meBART)
#' # Create a meBART object
#' # model <- mebart(x.train, y.train, ...)
#' # Create new data for prediction
#' # newdata <- matrix(...)
#' # Make predictions
#' #' predictions <- predict(model, newdata)
#' # Print predictions
#' #' print(predictions)

}
