% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.mebart_cont.R
\name{predict.mebart_cont}
\alias{predict.mebart_cont}
\title{Make Predictions for meBART Model with Continuous Outcomes}
\usage{
\method{predict}{mebart_cont}(object, newdata, mc.cores = 1, openmp = (mc.cores.openmp() > 0), ...)
}
\arguments{
\item{object}{A meBART object.}

\item{newdata}{A matrix of new data for prediction.}

\item{mc.cores}{Number of cores to use for parallel processing.}

\item{openmp}{Logical. If TRUE, OpenMP is used for parallel processing.}

\item{...}{Additional arguments.}
}
\value{
A matrix of predictions with size (ndpost x n)
}
\description{
This function makes predictions using a meBART model with \emph{continuous} outcomes.
}
\examples{
# Example usage:
set.seed(0)
x.train <- matrix(rnorm(1000), ncol = 10)
y.train <- rnorm(100)
mdl <- meBART::mebart(x.train, y.train,
              ndpost = 500,
              ntree = 50,
              meas_error_sigma = diag(10),
              x_mu = matrix(0, nrow=10, ncol=1),
              x_sigma = diag(10))
preds <- predict(mdl, newdata = x.train)
}
