% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.pbart.R
\name{predict.pbart}
\alias{predict.pbart}
\title{Make Predictions for meBART Model with Binary Outcomes}
\usage{
\method{predict}{pbart}(object, newdata, mc.cores = 1, openmp = (mc.cores.openmp() > 0), ...)
}
\arguments{
\item{object}{A meBART object.}

\item{newdata}{A matrix of new data for prediction.}

\item{mc.cores}{Number of cores to use for parallel processing.}

\item{openmp}{Logical. If TRUE, OpenMP is used for parallel processing.}

\item{...}{Additional arguments.}
}
\value{
A list with the following components:
\item{yhat.test}{A matrix of predictions in the latent Z space with size (ndpost x n).}
\item{prob.test}{A matrix of probabilities with size (ndpost x n).}
\item{prob.test.mean}{A vector of mean probabilities for each test sample.}
\item{binaryOffset}{The offset used in the model.}
}
\description{
This function makes predictions using a meBART model with \emph{binary} outcomes.
}
\examples{
# Example usage:
set.seed(0)
x.train <- matrix(rnorm(1000), ncol = 10)
y.train <- rbinom(100, 1, 0.5)
mdl <- meBART::pbart(x.train, y.train,
              ndpost = 500,
              ntree = 50,
              meas_error_sigma = diag(10),
              x_mu = matrix(0, nrow=10, ncol=1),
              x_sigma = diag(10))
preds <- predict(mdl, newdata = x.train)
}
