% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bartModelMatrix.R
\name{bartModelMatrix}
\alias{bartModelMatrix}
\title{BART Model Matrix}
\usage{
bartModelMatrix(
  X,
  numcut = 0L,
  usequants = FALSE,
  type = 7,
  rm.const = FALSE,
  cont = FALSE,
  xinfo = NULL
)
}
\arguments{
\item{X}{A data frame, matrix, or vector.}

\item{numcut}{The number of cut points for each variable.}

\item{usequants}{Logical. If TRUE, quantiles are used for cut points.}

\item{type}{The type of quantile to use.}

\item{rm.const}{Logical. If TRUE, constant variables are removed.}

\item{cont}{Logical. If TRUE, continuous variables are treated as such.}

\item{xinfo}{A list or matrix of cut points.}
}
\value{
A list containing the model matrix, number of cut points, removed variables, and cut points information.
}
\description{
This function creates a model matrix for BART.
}
\examples{
#' # Example usage
data(iris)
X <- iris[, -5]
numcut <- 3
usequants <- TRUE
rm.const <- TRUE
cont <- FALSE
xinfo <- NULL

#' result <- bartModelMatrix(X, numcut, usequants, rm.const, cont, xinfo)
#' print(result)

}
