% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.ind.R
\name{class.ind}
\alias{class.ind}
\title{#' @title Class Indicator Matrix
#' @description This function creates a class indicator matrix from a factor.}
\usage{
class.ind(cl)
}
\description{
#' @param cl A factor or vector of class labels.
}
\details{
#' @return A matrix with rows corresponding to observations and columns to classes.
}
