% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.ind.R
\name{class.ind}
\alias{class.ind}
\title{Class Indicator Matrix}
\usage{
class.ind(cl)
}
\arguments{
\item{cl}{A factor or vector of class labels.}
}
\value{
A (n x num_levels) matrix with rows corresponding to observations and columns corresponding to each unique level of the factor.
}
\description{
This function one hot encodes a factor to create a class indicator matrix.
}
\keyword{internal}
