% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.pmebart.R
\name{mc.pmebart}
\alias{mc.pmebart}
\title{Multi-core (parallel) meBART Predictions}
\usage{
mc.pmebart(
  x.test,
  treedraws,
  mu = 0,
  mc.cores = 2L,
  transposed = FALSE,
  dodraws = TRUE,
  nice = 19L
)
}
\arguments{
\item{x.test}{A matrix of test data.}

\item{treedraws}{A list containing the tree draws. (Part of the meBART model)}

\item{mu}{Offset for the predictions.}

\item{mc.cores}{Number of cores to use for parallel processing.}

\item{transposed}{Logical. Is x.test transposed?}

\item{dodraws}{Logical. If TRUE, all posterior draws are returned; if FALSE, only the mean is returned.}

\item{nice}{Nice level for the process.}
}
\value{
Prediction matrix of size (ndpost x n).
}
\description{
This function performs multi-core (parallel) predictions for meBART models.
}
\keyword{internal}
