% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meBART-package.R
\docType{package}
\name{meBART-package}
\alias{meBART}
\alias{meBART-package}
\title{meBART: Bayesian Additive Regression Trees with Measurement Error}
\description{
Bayesian Additive Regression Trees (BART) sum-of-trees model expanded to incorporate measurement errors in the independent variables.
}
\author{
\strong{Maintainer}: Kevin McCoy \email{kmm12@rice.edu}

Authors:
\itemize{
  \item Zachary Wooten
  \item Christine Peterson
}

}
\keyword{internal}
