% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meBART-package.R
\name{meBART}
\alias{meBART}
\title{meBART: Bayesian Additive Regression Trees with Measurement Error}
\description{
Bayesian Additive Regression Trees (BART) sum-of-trees model expanded to incorporate 
measurement errors in the independent variables.
}
