% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mebart.R
\name{mebart}
\alias{mebart}
\title{Fit a Measurement Error Bayesian Additive Regression Tree (meBART) Model}
\usage{
mebart(
  x.train,
  y.train,
  x.test = matrix(0, 0, 0),
  sparse = FALSE,
  theta = 0,
  omega = 1,
  a = 0.5,
  b = 1,
  augment = FALSE,
  rho = NULL,
  xinfo = matrix(0, 0, 0),
  usequants = FALSE,
  cont = FALSE,
  rm.const = TRUE,
  sigest = NA,
  sigdf = 3,
  sigquant = 0.9,
  k = 2,
  power = 2,
  base = 0.95,
  sigmaf = NA,
  lambda = NA,
  fmean = mean(y.train),
  w = rep(1, length(y.train)),
  ntree = 200L,
  numcut = 100L,
  ndpost = 1000L,
  nskip = 100L,
  keepevery = 1L,
  nkeeptrain = ndpost,
  nkeeptest = ndpost,
  nkeeptestmean = ndpost,
  nkeeptreedraws = ndpost,
  printevery = 100L,
  transposed = FALSE,
  proposal_sigma = meas_error_sigma,
  meas_error_sigma,
  x_mu,
  x_sigma
)
}
\arguments{
\item{x.train}{A matrix or data frame containing the explanatory variables for the training data.}

\item{y.train}{A numeric vector containing the continuous outcome variable for the training data.}

\item{x.test}{A matrix or data frame of test data (optional, default is an empty matrix). The structure should match `x.train`.}

\item{sparse}{A logical value (default is FALSE). If TRUE, uses a Dirichlet prior for variable selection, i.e. performs DART.}

\item{theta}{A numeric value (default is 0). alpha parameter in Linero 2018 equation (2), governs DART prior (DART ONLY)}

\item{omega}{A numeric value (default is 1). Unknown DART parameter. (DART ONLY)}

\item{a}{A numeric value (default is 0.5). Beta prior parameter. (DART ONLY)}

\item{b}{A numeric value (default is 1). Beta prior parameter. (DART ONLY)}

\item{augment}{A logical value (default is FALSE). FALSE uses Assumption 2.2 in Linero 2018, TRUE uses assumption 2.1 (DART ONLY)}

\item{rho}{A numeric value (default is NULL). Sparsity parameter used in DART. (DART ONLY)}

\item{xinfo}{A matrix of cutpoints (default is an empty matrix). Specifies cutpoints for the model if desired.}

\item{usequants}{A logical value (default is FALSE). If TRUE, quantiles are used for cutpoints; otherwise, uniform cutpoints are used.}

\item{cont}{A logical value (default is FALSE). Specifies if all variables are continuous.}

\item{rm.const}{A logical value (default is TRUE). If TRUE, removes constant columns from `x.train`.}

\item{sigest}{A numeric value (default is NA). Prior estimate for sigma; if NA, it is estimated from the data.}

\item{sigdf}{A numeric value (default is 3). Degrees of freedom for the sigma prior.}

\item{sigquant}{A numeric value (default is 0.90). Quantile for the sigma prior.}

\item{k}{A numeric value (default is 2.0). Prior parameter for leaf nodes.}

\item{power}{A numeric value (default is 2.0). Tree depth prior, beta in CGM18.}

\item{base}{A numeric value (default is 0.95). Tree depth prior, alpha in CGM18.}

\item{sigmaf}{A numeric value (default is NA). Parameter for the sigma prior, not scaled by the number of trees.}

\item{lambda}{A numeric value (default is NA). Scale parameter used in the sigma prior.}

\item{fmean}{A numeric value (default is the mean of `y.train`). The mean of the output variable `y.train`.}

\item{w}{A numeric vector (default is a vector of ones). Weights to multiply the standard deviation. Must be of length n.}

\item{ntree}{An integer (default is 200L). The number of trees to use in the model.}

\item{numcut}{An integer (default is 100L). The number of cutpoints to consider for each variable.}

\item{ndpost}{An integer (default is 1000L). The number of posterior draws to return.}

\item{nskip}{An integer (default is 100L). The number of observations for burn-in (MCMC).}

\item{keepevery}{An integer (default is 1L). The thinning parameter for MCMC.}

\item{nkeeptrain}{An integer (default is equal to `ndpost`). The number of training data posterior draws to keep.}

\item{nkeeptest}{An integer (default is equal to `ndpost`). The number of test data posterior draws to keep.}

\item{nkeeptestmean}{An integer (default is equal to `ndpost`). The number of MCMC iterations to return for the test mean.}

\item{nkeeptreedraws}{An integer (default is equal to `ndpost`). The number of MCMC iterations to return for tree draws.}

\item{printevery}{An integer (default is 100L). The number of iterations before progress is printed.}

\item{transposed}{A logical value (default is FALSE). Whether or not the input data is transposed.}

\item{proposal_sigma}{A numeric value (default is `meas_error_sigma`). Covariance matrix of the proposal distribution.}

\item{meas_error_sigma}{A numeric matrix. Covariance matrix of the measurement error, must be square and match the number of columns in `x.train`.}

\item{x_mu}{A numeric column vector. Mean of the explanatory variables.}

\item{x_sigma}{A numeric matrix. Covariance matrix of the explanatory variables.}
}
\value{
A list with the following elements:
\item{mu}{The mean of the outcome variable `y.train`, centered back in the predictions.}
\item{yhat.train.mean}{The mean of the posterior predictions for the training data.}
\item{yhat.train}{The posterior predictions for the training data, centered.}
\item{yhat.test.mean}{The mean of the posterior predictions for the test data.}
\item{yhat.test}{The posterior predictions for the test data, centered.}
\item{varcount}{A matrix of variable counts in the model.}
\item{varprob}{A matrix of variable probabilities in the model.}
\item{treedraws}{A list of posterior tree draws (if `nkeeptreedraws > 0`).}
\item{x_draws}{Tensor with posterior draws for the training data.}
\item{acceptances}{Whether or not the proposal was accepted.}
\item{proc.time}{The time taken for the C++ code to run.}
\item{varcount.mean}{The mean of the variable counts.}
\item{varprob.mean}{The mean of the variable probabilities.}
\item{rm.const}{A vector specifying which columns were removed from `x.train`.}
}
\description{
This function fits a meBART model with continuous output data.
}
\examples{
# Example usage:
set.seed(0)
x.train <- matrix(rnorm(1000), ncol = 10)
y.train <- rnorm(100)
mdl <- mebart(x.train, y.train,
              ndpost = 500,
              ntree = 50,
              meas_error_sigma = diag(10),
              x_mu = matrix(0, nrow=10, ncol=1),
              x_sigma = diag(10))
preds <- mdl$yhat.train.mean
print(paste0("Training MSE: ", round(mean((y.train - preds)^2), 3)))
}
