% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmebart.R
\name{pmebart}
\alias{pmebart}
\title{Single-core + Multi-core (OpenMP) meBART Predictions}
\usage{
pmebart(
  x.test,
  treedraws,
  mu = 0,
  mc.cores = 1L,
  transposed = FALSE,
  dodraws = TRUE,
  nice = 19L
)
}
\arguments{
\item{x.test}{A matrix of test data.}

\item{treedraws}{A list containing the tree draws. (Part of the meBART model)}

\item{mu}{Offset for the predictions.}

\item{mc.cores}{Number of cores to use for parallel processing.}

\item{transposed}{Logical. Is x.test transposed?}

\item{dodraws}{Logical. If TRUE, all posterior draws are returned; if FALSE, only the mean is returned.}

\item{nice}{Nice level for the process. I'm pretty sure this is not used in the code.}
}
\value{
Prediction matrix of size (ndpost x n).
}
\description{
This function performs single-core or multi-core (OpenMP) predictions for meBART 
models. This function is also called by `mc.pmebart.R`.
}
\keyword{internal}
